add moshi custom adapter for rate

+++ app/src/main/java/eu/szwiec/checkittravelkit/repository/data/Rate.kt
@@ -1,7 +1,7 @@
 package eu.szwiec.checkittravelkit.repository.data
 
 data class Rate(
-        val value: Float = 0.0f,
+        val value: Double = 0.0,
         val fromCurrencyCode: String = "",
         val lastUpdate: Long = 0
 )
